Refactor: Split classes up into separate files
Also, split up the Minimap.init method into smaller methods to make it a bit more readable
